snart er vi klar til at integrere shortprod x64 version, som bl.a. skal bruger den nye klasse _int128. Mangler stadig test af hele molevitten

+++ Lib/Src/BigReal/BigRealThreadPool.cpp
@@ -78,7 +78,7 @@ void BigRealThreadPool::releaseMTThreadArray(MThreadArray &threads) { // static
 
       instance.m_MTThreadPool.releaseResource(thread);
     }
-    instance.m_activeThreads -= threads.size();
+    instance.m_activeThreads -= (int)threads.size();
     threads.clear();
   }
   instance.m_gate.signal();
